Expanded testing

+++ test/setup.ts
@@ -1,14 +1,28 @@
 import "mocha";
 import { setup } from "./constants";
 import mongoose from "mongoose";
+import { beforeEach } from "mocha";
 
-// clear database after connected
+export function clear() {
+  shouldClear = true;
+}
+
+let shouldClear = true;
+
 before(function (done) {
   this.timeout(15000);
   const db = mongoose.connection;
   db.once("open", async () => {
-    await db.dropDatabase();
-    await setup();
     done();
   });
+});
+
+// clear database before each test
+beforeEach(async () => {
+  if (shouldClear) {
+    const db = mongoose.connection;
+    await db.dropDatabase();
+    await setup();
+    shouldClear = false;
+  }
 });
